First version of Q2
I train only a small model so far. I will check the code to make sure everything is ok in the next few days and do a bigger model to reduce the Elbo. Then I will answer the second part of the question. I will try to finish the question for Thursday.